refactor(remote-storage): separate thread pool and quota configurations
Split remote log manager dynamic configurations into two separate handlers:
- RemoteLogDynamicThreadPool for thread pool size configs
- DynamicRemoteLogConfig for quota and cache configs

Changes:
- Move thread pool configs from DynamicRemoteLogConfig to new RemoteLogDynamicThreadPool class
- Add thread pool configs to DynamicThreadPool.ReconfigurableConfigs
- Change thread pool size parameters from long to int
- Add thread pool size getters to KafkaConfig
- Add resize method to RemoteStorageThreadPool

This separation improves code organization by grouping related configs together
and fixes type consistency issues with thread pool sizes.

+++ core/src/main/java/kafka/log/remote/RemoteLogManager.java
@@ -289,17 +289,17 @@ public class RemoteLogManager implements Closeable {
         rlmFetchQuotaManager.updateQuota(new Quota(quota, true));
     }
 
-    public void updateCopyThreadPoolSize(long size) {
+    public void updateCopyThreadPoolSize(int size) {
         LOGGER.info("Updating remote copy thread pool size to {}", size);
         rlmCopyThreadPool.resize(size);
     }
 
-    public void updateExpirationThreadPoolSize(long size) {
+    public void updateExpirationThreadPoolSize(int size) {
         LOGGER.info("Updating remote expiration thread pool size to {}", size);
         rlmExpirationThreadPool.resize(size);
     }
 
-    public void updateReaderThreadPoolSize(long size) {
+    public void updateReaderThreadPoolSize(int size) {
         LOGGER.info("Updating remote reader thread pool size to {}", size);
         remoteStorageReaderThreadPool.resize(size);
     }
@@ -2178,8 +2178,8 @@ public class RemoteLogManager implements Closeable {
             scheduledThreadPool = createPool();
         }
 
-        public void resize(long newSize) {
-            scheduledThreadPool.setCorePoolSize((int) newSize);
+        public void resize(int newSize) {
+            scheduledThreadPool.setCorePoolSize(newSize);
         }
 
         private ScheduledThreadPoolExecutor createPool() {
